renamed components, refactored carousel from stateful to stateless, and added css modules
