+++ client/src/components/Carousel.css
@@ -1,0 +1,21 @@
+.carouselContainer {
+    width: 3000px;
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    grid-template-rows: 50% 50%;
+    grid-auto-flow: row; 
+    grid-gap: 2px;
+}
+
+.firstImageContainer {
+    grid-column: span 2;
+    grid-row: span 2;
+    background-size: cover;
+    width: 100%;
+    overflow: hidden;
+}
+
+.imageContainer {
+    width: 100%;
+    height: 100%;
+}
