+++ client/src/components/LeftArrow.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import leftArrowStyle from './LeftArrow.css';
 
 const LeftArrow = props => (
-  <span className="leftArrowBlock" onClick={props.goBack}>
+  <span className={leftArrowStyle.leftArrowBlock} onClick={props.goBack}>
     <i className="fas fa-angle-left" />
   </span>
 );
