+++ client/src/components/RightArrow.css
@@ -1,0 +1,30 @@
+.fas.fa-angle-right {
+    font-size: 45px;
+    float: right;
+    position: relative;
+    display: inline-block;
+    top: 0px;
+    color: #fff8ea;
+  }
+  
+  .rightArrowBlock {
+    opacity: 0.7;
+    float: right;
+    position: relative;
+    display: inline-block;
+    top: 205px;
+    z-index: 500;
+    background-color: #4f4849;
+    margin-right: 5px;
+    border: solid white 0.5px;
+    padding: 5px;
+  }
+  
+  .rightArrowBlock:hover {
+    opacity: 0.95;
+    transition: 0.3s;
+  }
+  
+  .fas.fa-angle-right:hover {
+    color: #4198FF;
+  }
