+++ client/src/components/SlideShow.css
@@ -1,0 +1,4 @@
+.selectedImage {
+    position: fixed;
+    left: 15%;
+}
