+++ client/src/components/SlideShow.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import slideShow from './SlideShow.css';
+
+const SlideShow = (props) => (
+  <div onClick={props.renderImage}>
+    <img src={`${props.image}`} className={slideShow.selectedImage} />
+  </div>
+);
+
+export default SlideShow;
