+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Carousel from './components/Carousel.jsx';
+import App from './components/App.jsx';
 
-ReactDOM.render(<Carousel />, document.getElementById('carousel'));
+ReactDOM.render(<App />, document.getElementById('carousel'));
