+++ webpack.config.js
@@ -12,13 +12,13 @@ module.exports = {
       },
     }, {
       test: /\.css$/,
-      loader: 'style-loader'
+      loader: 'style-loader',
     }, {
       test: /\.css$/,
       loader: 'css-loader',
       query: {
         modules: true,
-        localIdentName: '[name]__[local]___[hash:base64:5]'
+        localIdentName: '[name]__[local]___[hash:base64:5]',
       },
     }, {
       test: /\.js$/,
